pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Hexlet project difference calculator"
authors = ["Madixxx22 <vysokolyan2016@gmail.com>"]
readme = "README.md"
packages = [{include = "gendiff"}]
repository = "https://github.com/Madixxx22/python-project-50"
classifiers = ["Topic :: Software Development :: :Libraries :: Python Modules"]

[tool.poetry.dependencies]
python = "^3.10"
argparse = "^1.4.0"
pytest = "^7.2.0"
flake8 = "^6.0.0"
pytest-cov = "^4.0.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"